Update the packages
I actually just wiped the package.json and installed the newest versions of
everything. I figured that was probably the easiest way to delay the next time
I need to deal with a depedabot PR for the longest.

I did install the main dependencies with --no-optional, but I'm pretty sure
they still installed the graphql dependencies, even though nest is now nest7,
so, yeah. Hmm.

This leaves a whole bunch of stuff not working that I will fix in the next
commit.