Mass package update (#26)
* Update readme with slightly more specific details

With regards to actually getting the thing running locally. Could do with being
even more fleshed out, but I think this is at least better than it was.

* Update the packages

I actually just wiped the package.json and installed the newest versions of
everything. I figured that was probably the easiest way to delay the next time
I need to deal with a depedabot PR for the longest.

I did install the main dependencies with --no-optional, but I'm pretty sure
they still installed the graphql dependencies, even though nest is now nest7,
so, yeah. Hmm.

This leaves a whole bunch of stuff not working that I will fix in the next
commit.

* Update the codebase for new packages

Largely just name-changing find and replace level stuff, but class-transformer
did also change the signature of the @Transform decorator, so I had to deal
with that too. Pretty sure I got the equivalent of what we had.